TOR management is based on OVS protocol. A new role called TSN (TOR services node)
is introduced for compute node. TSN handles multicast / broadcast traffic from TOR,
along with DHCP and DNS services. A new process contrail-tor-agent will be run for
every TOR that is managed. This process handles configuring TOR using OVSDB, as also
exporting the routes reported by the TOR.

Gateway Port on compute node can be configured with a subnet that is reachable on
the port.

+++ src/vnsw/agent/test/test_xmpp_discovery.cc
@@ -254,7 +254,8 @@ TEST_F(AgentXmppUnitTest, XmppConnection_Discovery) {
     client->WaitForIdle();
 
     //wait for connection establishment
-    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(0) == xs1->GetPort()); 
+    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(0) ==
+                (uint32_t)xs1->GetPort());
     WAIT_FOR(1000, 10000, 
         agent_->controller_xmpp_channel(0)->GetXmppChannel()->GetPeerState()
         == xmps::READY); 
@@ -276,7 +277,8 @@ TEST_F(AgentXmppUnitTest, XmppConnection_Discovery) {
     client->WaitForIdle();
 
     //wait for connection establishment
-    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(1) == xs2->GetPort()); 
+    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(1) ==
+                (uint32_t)xs2->GetPort());
     WAIT_FOR(1000, 10000, 
         agent_->controller_xmpp_channel(1)->GetXmppChannel()->GetPeerState() 
         == xmps::READY); 
@@ -298,7 +300,8 @@ TEST_F(AgentXmppUnitTest, XmppConnection_Discovery) {
     client->WaitForIdle();
 
     //wait for connection establishment
-    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(0) == xs3->GetPort()); 
+    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(0) ==
+                (uint32_t)xs3->GetPort());
     WAIT_FOR(1000, 10000, 
         agent_->controller_xmpp_channel(0)->GetXmppChannel()->GetPeerState() 
         == xmps::READY); 
@@ -318,12 +321,14 @@ TEST_F(AgentXmppUnitTest, XmppConnection_Discovery) {
     client->WaitForIdle();
 
     //wait for connection establishment
-    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(0) == xs3->GetPort()); 
+    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(0) ==
+                (uint32_t)xs3->GetPort());
     WAIT_FOR(1000, 10000, 
         agent_->controller_xmpp_channel(0)->GetXmppChannel()->GetPeerState() 
         == xmps::READY); 
 
-    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(1) == xs4->GetPort()); 
+    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(1) ==
+                (uint32_t)xs4->GetPort());
     WAIT_FOR(1000, 10000, 
         agent_->controller_xmpp_channel(1)->GetXmppChannel()->GetPeerState() 
         == xmps::READY); 
@@ -349,12 +354,14 @@ TEST_F(AgentXmppUnitTest, XmppConnection_Discovery) {
     client->WaitForIdle();
 
     //wait for connection establishment
-    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(0) == xs5->GetPort()); 
+    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(0) ==
+                (uint32_t)xs5->GetPort());
     WAIT_FOR(1000, 10000, 
         agent_->controller_xmpp_channel(0)->GetXmppChannel()->GetPeerState() 
         == xmps::READY); 
 
-    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(1) == xs6->GetPort()); 
+    EXPECT_TRUE(agent_->controller_ifmap_xmpp_port(1) ==
+                (uint32_t)xs6->GetPort());
     WAIT_FOR(1000, 10000, 
         agent_->controller_xmpp_channel(1)->GetXmppChannel()->GetPeerState() 
         == xmps::READY);
